fix(synthstrip): shape calculation needs reordering
The calculation of the new shape within the preprocessing for synthstrip
was done in XYZ coordinates (i.e., extents for each physical axis were
estimated and divided by the new 1mm pixel size).
However, the new axis sizes did not account for the fact that data was
stored with IJK -> XZY convention (corresponding to LIA).

This PR accounts for this miscalculation.

Resolves: #999.