enh: drop FSL MELODIC (without alternative)
After careful consideration, I'm proposing to drop the ``--ica``
argument of the functional workflow, to continue toward eliminating
FSL as a dependency.

The reason is that the ICA decomposition does not seem (alone)
sufficient to assess the quality.
Or at least, the kind of problems/artifacts you can pick with just the
ICA decomposition (without assistance like CompCor, AROMA or FIX) are
typically identifiable with the standard deviation image and other
mosaics.

I would not be against adding sklearn's FastICA or a reimplementation
of MELODIC to have the ICA + EM mixtrue modelling, but, for now, I
believe the priority should be to finally remove FSL, even if that means
discontinuing a feature.

Other than the visual reports, there's no other side effect since we are
currently not calculating any IQMs from the ICA results (which would be
a very interesting addition in the future, perhaps with some AROMA
assistance).

Relates: #1032.