BUG: Fix elapsed runtime format
Fix elapsed runtime format: the `strftime` function `%H` format
specifier only supports 0–23 hours (for a single day), thus, when the
time exceeds 24 hours, the hours component resets to 0.

This patch set computes the elapsed time (per participant and for the
entire run), builds a `datetime.timedelta` instance and formats the
elapsed time following the `%Hh %Mmin %Ss` formatting using a function
created for that purpose.

Do not call `time.gmtime` as when reporting the elapsed time we are not
interested in the UTC time format.