ENH: Remove additional `setup-python` action in GHA workflow file
Remove additional `setup-python` action in GHA workflow file, as Python
is available through `pipx`.

Fixes:
```
ruff
The `python-version` input is not set. The version of Python currently in `PATH` will be used.
```

raised for example at:
https://github.com/nipreps/mriqc/actions/runs/12795005155

Also, revert the change introduced in commit 5cf7010 that unsuccessfully
intended to fix the warning.